feat(engagement): E1 streak hype panel — StreakHypeCard wired into feed (build 164)
Adds StreakHypeCard as the first SliverToBoxAdapter above the feed on the
home screen. The card reads currentStreak / longestStreak from the user doc
in real-time and renders a gradient fire-icon banner when streak > 0.

Tier thresholds: 1–6 days → green, 7–13 → gold, 14–29 → steel-blue,
30+ → champion purple/gold. Tap opens StreakHistorySheet with milestone
progress bar (3 / 7 / 14 / 30 days). Session-dismissible (not persisted).

Also includes dart fix --apply quote normalisation across 86 service /
widget files (single-quote style enforcement per analysis_options.yaml).
No logic changes in any normalised file; voice_controller.dart audio
session config is untouched.

+++ lib/core/services/weekly_summary_service.dart
@@ -149,7 +149,7 @@ class WeeklySummaryService {
       totalDurationSeconds: totalDuration,
       totalCalories: totalCalories,
       totalElevationGain: totalElevation,
-      avgPaceSecondsPerKm: avgPace.toDouble(),
+      avgPaceSecondsPerKm: avgPace,
       longestRunKm: longestRun,
       fastestPaceSecondsPerKm: fastestPace == double.infinity ? 0 : fastestPace,
     );
